Refactor FeaturesModel
FeaturesModel is now a base class, inherited by:
LayerFeaturesModel (previously FeaturesModel)
and StaticFeaturesModel (not tied to layer - no fetching)
