+++ app/test/testmodels.h
@@ -22,8 +22,8 @@ class TestModels : public QObject
     void init(); // will be called before each testfunction is executed.
     void cleanup(); // will be called after every testfunction.
 
-    void testFeaturesModel();
-    void testFeaturesModelSorted();
+    void testLayerFeaturesModel();
+    void testLayerFeaturesModelSorted();
     void testValueRelationFeaturesModel();
     void testProjectsModel();
     void testProjectsProxyModel();
