+++ app/valuerelationfeaturesmodel.h
@@ -11,7 +11,7 @@
 #define VALUERELATIONFEATURESMODEL_H
 
 #include "inputconfig.h"
-#include "featuresmodel.h"
+#include "layerfeaturesmodel.h"
 #include "featurelayerpair.h"
 
 #include <QObject>
@@ -21,7 +21,7 @@
  * ValueRelationFeaturesModel class lists features from a specific layer regarding to a filterExpression of value relations.
  * It is used as a model in ValueRelations QML editors.
  */
-class ValueRelationFeaturesModel : public FeaturesModel
+class ValueRelationFeaturesModel : public LayerFeaturesModel
 {
     Q_OBJECT
